% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-beta_matrix.R
\name{beta_matrix}
\alias{beta_matrix}
\title{Beta value matrix}
\format{
A tibble with 20 rows and 3 variables
\describe{
\item{marker1}{Beta values of marker1 for all samples}
\item{marker2}{Beta values of marker2 for all samples}
\item{marker3}{Beta values of marker3 for all samples}
}
}
\usage{
data("beta_matrix")
}
\value{
A tibble with 20 rows and 3 variables
}
\description{
A list of methylation levels (e.g., beta values), 
where each row is a sample and each column is a marker
}
\author{
Ran Hu \email{huran@ucla.edu}
}
