% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-CancerDetector.markers.R
\name{CancerDetector.markers}
\alias{CancerDetector.markers}
\title{Cancer-specific marker parameter}
\format{
A tibble with 1266 rows and 3 variables
\describe{
\item{markerName}{Name of the marker}
\item{tumor}{Paired beta distribution shape parameters 
for tumor samples}
\item{normalPlasma}{Paired beta distribution shape 
parameters for normal plasma samples}
}
}
\usage{
data("CancerDetector.markers")
}
\value{
A tibble with 1266 rows and 3 variables
}
\description{
The paired shape parameters of beta distributions 
for cancer-specific markers
}
\author{
Ran Hu \email{huran@ucla.edu}
}
