% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cfDeconvolve.markers.R
\name{cfDeconvolve.markers}
\alias{cfDeconvolve.markers}
\title{Tissue-specific marker parameter}
\format{
A tibble with 10 rows and 8 variables
\describe{
\item{markerName}{Name of the marker}
\item{tissue1}{Paired beta distribution shape parameters 
for tissue1 samples}
\item{tissue2}{Paired beta distribution shape parameters 
for tissue2 samples}
\item{tissue3}{Paired beta distribution shape parameters 
for tissue3 samples}
\item{tissue4}{Paired beta distribution shape parameters 
for tissue4 samples}
\item{tissue5}{Paired beta distribution shape parameters 
for tissue5 samples}
\item{tissue6}{Paired beta distribution shape parameters 
for tissue6 samples}
\item{tissue7}{Paired beta distribution shape parameters 
for tissue7 samples}
}
}
\usage{
data("cfDeconvolve.markers")
}
\value{
A tibble with 10 rows and 8 variables
}
\description{
The paired shape parameters of beta distributions 
for tissue-specific markers
}
\author{
Ran Hu \email{huran@ucla.edu}
}
