% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfSort.R
\name{cfSort}
\alias{cfSort}
\title{cfSort: tissue deconvolution}
\usage{
cfSort(readsBinningFile, id = "sample")
}
\arguments{
\item{readsBinningFile}{a file of the fragment-level methylation states 
of reads that mapped to the cfSort markers. In compressed form.}

\item{id}{the sample ID.}
}
\value{
the tissue composition of the cfDNA sample.
}
\description{
Tissue deconvolution in cfDNA using DNN models.
}
\examples{
## input files
demo.dir <- system.file("data", package="cfTools")
readsBinningFile <- file.path(demo.dir, "cfsort_reads.txt.gz")
id <- "test"

cfSort(readsBinningFile, id)

}
