% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cfsort_markers.R
\name{cfsort_markers}
\alias{cfsort_markers}
\title{cfSort markers}
\format{
A tibble with 51035 rows and 4 variables
\describe{
\item{marker_index}{The marker index used in cfSort method}
\item{alpha_threshold}{The alpha threshold for each marker}
\item{pair}{The pair of tissues used for identifying the marker}
\item{group}{The group number for each marker}
}
}
\usage{
data("cfsort_markers")
}
\value{
A tibble with 51035 rows and 4 variables
}
\description{
Marker information for the cfSort function, 
where each row is the information about a marker
}
\author{
Ran Hu \email{huran@ucla.edu}
}
