% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-CpG_OB_demo.R
\name{CpG_OB_demo}
\alias{CpG_OB_demo}
\title{Methylation information for CpG on the original bottom strand (OB)}
\format{
A tibble with 2224 rows and 5 variables
\describe{
\item{sequence ID}{ID of the sequence}
\item{methylation state}{Methylated or unmethylated CpG site}
\item{chromosome name}{Chromosome name}
\item{chromosome start}{Chromosome start position}
\item{methylation call}{Methylation call}
}
}
\usage{
data("CpG_OB_demo")
}
\value{
A tibble with 2224 rows and 5 variables
}
\description{
Methylation information for CpG on the original 
bottom strand (OB), which is one of the outputs from 
'bismark methylation extractor'
}
\author{
Ran Hu \email{huran@ucla.edu}
}
