% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.refo_frag.bed.R
\name{demo.refo_frag.bed}
\alias{demo.refo_frag.bed}
\title{Fragment-level information}
\format{
A tibble with 559 rows and 6 variables
\describe{
\item{chr}{Chromosome}
\item{start}{Chromosome start}
\item{end}{Chromosome end}
\item{fragmentLength}{Fragment length}
\item{strand}{Strand}
\item{name}{ID of the sequence}
}
}
\usage{
data("demo.refo_frag.bed")
}
\value{
A tibble with 559 rows and 6 variables
}
\description{
A BED file of fragment-level information
}
\author{
Ran Hu \email{huran@ucla.edu}
}
