% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.refo_meth.bed.R
\name{demo.refo_meth.bed}
\alias{demo.refo_meth.bed}
\title{Methylation information on fragments}
\format{
A tibble with 552 rows and 8 variables
\describe{
\item{chr}{Chromosome}
\item{cpgStart}{Start postion of first CpG on the fragment}
\item{cpgEnd}{End postion of first CpG on the fragment}
\item{strand}{Strand}
\item{cpgNumber}{Number of CpG sites on the fragment}
\item{cpgPosition}{Postions of CpG sites on the fragment}
\item{methState}{A string of methylation states of 
CpG sites on the fragment}
\item{name}{ID of the sequence}
}
}
\usage{
data("demo.refo_meth.bed")
}
\value{
A tibble with 552 rows and 8 variables
}
\description{
A BED file of methylation information on fragments
}
\author{
Ran Hu \email{huran@ucla.edu}
}
