% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.sorted.bed.R
\name{demo.sorted.bed}
\alias{demo.sorted.bed}
\title{Paired-end sequencing reads}
\format{
A tibble with 1117 rows and 6 variables
\describe{
\item{chr}{Chromosome name}
\item{start}{Chromosome start}
\item{end}{Chromosome end}
\item{name}{Sequence ID}
\item{score}{Mapping quality score}
\item{strand}{Strand}
}
}
\usage{
data("demo.sorted.bed")
}
\value{
A tibble with 1117 rows and 6 variables
}
\description{
Paired-end sequencing reads information
}
\author{
Ran Hu \email{huran@ucla.edu}
}
