% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFragMeth.R
\name{GenerateFragMeth}
\alias{GenerateFragMeth}
\title{Generate fragment level methylation status}
\usage{
GenerateFragMeth(
  frag_bed,
  meth_bed,
  output.dir = "",
  id = "",
  python = "python"
)
}
\arguments{
\item{frag_bed}{a BED file containing information for every read, which is the output of CollapsePEReads().}

\item{meth_bed}{a BED file containing methylation status on every read, which is the output of CollapseCpGs().}

\item{output.dir}{a path to the output directory. Default is "", which means the output will not be written into a file.}

\item{id}{an ID name for the input data. Default is "", which means the output will not be written into a file.}

\item{python}{a path to Python 3. Default is "python".}
}
\value{
a data frame in BED file format and/or written to an output BED file.
}
\description{
Join two data frames containing read information and methylation status on each read into one data frame.
}
\examples{
## input files
demo.dir <- system.file("extdata", package="cfTools")
frag_bed <- read.delim(file.path(demo.dir, "demo.refo_frag.bed"))
meth_bed <- read.delim(file.path(demo.dir, "demo.refo_meth.bed"))

output <- GenerateFragMeth(frag_bed, meth_bed)

}
