% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_marker_file.R
\name{generateMarkerFile}
\alias{generateMarkerFile}
\title{Generate the methylation pattern of markers}
\usage{
generateMarkerFile(x, sample.types, marker.indexes, output.file = "")
}
\arguments{
\item{x}{a data frame of beta values, where each row is a sample and each column is a marker.}

\item{sample.types}{a vector of sample types corresponding to the rows of the data frame.}

\item{marker.indexes}{a vector of marker indexes corresponding to the columns of the data frame.}

\item{output.file}{a character string naming the output file. Default is "", which means the output will not be written into a file.}
}
\value{
a data frame containing the paired shape parameters of a beta distribution for markers and/or written to an output file.
}
\description{
Output paired shape parameters of a beta distribution for markers.
}
\examples{
## input files
demo.dir <- system.file("extdata", package="cfTools")
x <- read.csv(file.path(demo.dir, "beta_matrix.csv"), row.names=1)
sample.types <- read.csv(file.path(demo.dir, "sample_type.csv"), row.names=1)$Sample.Type
marker.indexes <- read.csv(file.path(demo.dir, "marker_index.csv"), row.names=1)$Marker.Index

output <- generateMarkerFile(x, sample.types, marker.indexes)

}
