% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-markers_index.R
\name{marker_index}
\alias{marker_index}
\title{Marker name}
\format{
A tibble with 3 rows and 1 variables
\describe{
\item{markerIndex}{Marker name}
}
}
\usage{
data("marker_index")
}
\value{
A tibble with 3 rows and 1 variables
}
\description{
A vector of marker names corresponding to the columns 
of the list of methylation levels.
}
\author{
Ran Hu \email{huran@ucla.edu}
}
