% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-markers.bed.R
\name{markers.bed}
\alias{markers.bed}
\title{Genomic postions of markers}
\format{
A tibble with 3 rows and 4 variables
\describe{
\item{chr}{Chromosome}
\item{start}{Chromosome start}
\item{end}{Chromosome end}
\item{markerName}{Marker name}
}
}
\usage{
data("markers.bed")
}
\value{
A tibble with 3 rows and 4 variables
}
\description{
A BED file of genomic regions of markers
}
\author{
Ran Hu \email{huran@ucla.edu}
}
