% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample_type.R
\name{sample_type}
\alias{sample_type}
\title{Sample type}
\format{
A tibble with 20 rows and 1 variables
\describe{
\item{sampleType}{Sample type}
}
}
\usage{
data("sample_type")
}
\value{
A tibble with 20 rows and 1 variables
}
\description{
A vector of sample types (e.g., tumor or normal, 
tissue types) corresponding to the rows of the 
list of methylation levels.
}
\author{
Ran Hu \email{huran@ucla.edu}
}
