% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerDetector.R
\name{CancerDetector}
\alias{CancerDetector}
\title{Cancer Detector}
\usage{
CancerDetector(readsBinningFile, tissueMarkersFile, python = "python3")
}
\arguments{
\item{readsBinningFile}{a file of the fragment-level methylation states of reads that mapped to the markers.}

\item{tissueMarkersFile}{a file of paired shape parameters of beta distributions for markers.}

\item{python}{a path to Python 3. Default is "python3".}
}
\value{
cfDNA tumor burden and normal cfDNA fraction.
}
\description{
Detect tumor-derived cfDNA and estimate the tumor burden.
}
\examples{
## input files
demo.dir <- system.file("extdata", package="cfTools")
readsBinningFile <- file.path(demo.dir, "CancerDetector.reads.txt")
tissueMarkersFile <- file.path(demo.dir, "CancerDetector.markers.txt")

CancerDetector(readsBinningFile, tissueMarkersFile)

}
