% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfSort.R
\name{cfSort}
\alias{cfSort}
\title{cfDNA Methylation Read Deconvolution}
\usage{
cfSort(
  readsBinningFile,
  tissueMarkersFile,
  outputType = "tissueFraction",
  outputFile,
  numTissues,
  emAlgorithmType = "em.global.unknown",
  likelihoodRatioThreshold = 2,
  emMaxIterations = 100
)
}
\arguments{
\item{readsBinningFile}{input cfDNA methylation sequencing reads file, either in plain text or compressed form.}

\item{tissueMarkersFile}{input markers file in plain text. Each marker for each tissue type has either a median beta value or a paired shape parameters of beta-distribution from the tissue population of tissue types.}

\item{outputType}{output type: tissueFraction (default), tissueFraction+readCountRaw, tissueFraction+readCountPerMillion, tissueFraction+readCountPerBillion.}

\item{outputFile}{output file.}

\item{numTissues}{number of tissue types.}

\item{emAlgorithmType}{read-based tissue deconvolution EM algorithm type: em.global.unknown (default), em.global.known, em.local.unknown, em.local.known.}

\item{likelihoodRatioThreshold}{a positive float number. We suggest this number is 2 (default). All reads with likelihood ratio < cutoff (default: -1.0, meanin no reads filtering is used) will not be used. Likelihood ratio is the max(all tissues' likelihoods)/min(all tissues' likelihoods).}

\item{emMaxIterations}{EM algorithm maximum iteration number. Default is 100.}
}
\value{
an output file.
}
\description{
Read deconvolution.
}
\examples{
## input files
demo.dir <- system.file("extdata", package="cfTools")
readsBinningFile <- file.path(demo.dir, "ReadDeconvolution.reads.txt")
tissueMarkersFile <- file.path(demo.dir, "ReadDeconvolution.markers.txt")
outputType <- "tissueFraction+readCountPerBillion"
outputFile <- file.path(demo.dir, "example.profile")
numTissues <- 7
emAlgorithmType <- "em.global.unknown"
likelihoodRatioThreshold <- 2

cfSort(readsBinningFile, tissueMarkersFile, outputType, outputFile,
numTissues, emAlgorithmType, likelihoodRatioThreshold)

}
